+++ tests/unit-tests/Ares/Schema/ParserTest.php
@@ -20,7 +20,6 @@ use Ares\Schema\SchemaMap;
 use Ares\Schema\SchemaReference;
 use Ares\Schema\SchemaTuple;
 use Ares\Schema\TypeRegistry;
-use Ares\Validation\RuleFactory;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -31,7 +30,6 @@ use PHPUnit\Framework\TestCase;
 class ParserTest extends TestCase
 {
     /**
-     * @covers ::__construct
      * @covers ::ascertainInputHoldsArrayOrFail
      * @covers ::extractTypeOrFail
      * @covers ::fail
@@ -45,7 +43,6 @@ class ParserTest extends TestCase
      *
      * @dataProvider getParseErrorHandlingSamples
      *
-     * @param RuleFactory $ruleFactory              Validation rule factory.
      * @param mixed       $schemaIn                 Provided schema.
      * @param array|null  $expectedSchemaOut        Expected resulting schema.
      * @param string|null $expectedException        Expected exception.
@@ -53,14 +50,12 @@ class ParserTest extends TestCase
      * @return void
      */
     public function testParseErrorHandling(
-        RuleFactory $ruleFactory,
         $schemaIn,
         ?array $expectedSchemaOut = null,
         ?string $expectedException = null,
         ?string $expectedExceptionMessage = null
     ): void {
-        //$ruleFactory = new RuleFactory();
-        $parser = new Parser($ruleFactory);
+        $parser = new Parser();
 
         if ($expectedException !== null) {
             $this->expectException($expectedException);
@@ -84,14 +79,12 @@ class ParserTest extends TestCase
     {
         return [
             'invalid schema' => [
-                new RuleFactory(),
                 'foobar',
                 null,
                 InvalidSchemaException::class,
                 'Invalid schema value:  must be of type <array>, got <string>',
             ],
             'multiple types' => [
-                new RuleFactory(),
                 [
                     'type' => 'integer',
                     ['type' => 'integer'],
@@ -101,14 +94,12 @@ class ParserTest extends TestCase
                 'Ambiguous schema:  contains multiple `type` validation rules',
             ],
             'missing type' => [
-                new RuleFactory(),
                 [],
                 null,
                 InvalidSchemaException::class,
                 'Insufficient schema:  contains no `type` validation rule',
             ],
             'invalid type' => [
-                new RuleFactory(),
                 [
                     'type' => 'foo',
                 ],
@@ -117,7 +108,6 @@ class ParserTest extends TestCase
                 'Invalid schema:  uses unknown type: "foo"',
             ],
             'invalid schema value' => [
-                new RuleFactory(),
                 [
                     'type' => 'string',
                     'foobar',
@@ -127,7 +117,6 @@ class ParserTest extends TestCase
                 'Invalid schema value: /0 must be of type <array>, got <string>',
             ],
             'type "list" without "schema"' => [
-                new RuleFactory(),
                 [
                     'type' => 'list',
                 ],
@@ -136,7 +125,6 @@ class ParserTest extends TestCase
                 'Missing schema key:  uses type "list" but contains no "schema" key',
             ],
             'nested type "list" without "schema"' => [
-                new RuleFactory(),
                 [
                     'type' => 'list',
                     'schema' => [
@@ -148,7 +136,6 @@ class ParserTest extends TestCase
                 'Missing schema key: /schema uses type "list" but contains no "schema" key',
             ],
             'type "map" without "schema"' => [
-                new RuleFactory(),
                 [
                     'type' => 'map',
                 ],
@@ -157,7 +144,6 @@ class ParserTest extends TestCase
                 'Missing schema key:  uses type "map" but contains no "schema" key',
             ],
             'nested type "map" without "schema"' => [
-                new RuleFactory(),
                 [
                     'type' => 'map',
                     'schema' => [
@@ -171,7 +157,6 @@ class ParserTest extends TestCase
                 'Missing schema key: /schema/name uses type "map" but contains no "schema" key',
             ],
             'type "list" with non-array schema value' => [
-                new RuleFactory(),
                 [
                     'type' => 'list',
                     'schema' => 'foo',
@@ -181,7 +166,6 @@ class ParserTest extends TestCase
                 'Invalid schema value: /schema must be of type <array>, got <string>',
             ],
             'nested type "list" with non-array schema value' => [
-                new RuleFactory(),
                 [
                     'type' => 'list',
                     'schema' => [
@@ -194,7 +178,6 @@ class ParserTest extends TestCase
                 'Invalid schema value: /schema/schema must be of type <array>, got <double>',
             ],
             'type "map" with non-array schema value' => [
-                new RuleFactory(),
                 [
                     'type' => 'map',
                     'schema' => 'foo',
@@ -204,7 +187,6 @@ class ParserTest extends TestCase
                 'Invalid schema value: /schema must be of type <array>, got <string>',
             ],
             'nested type "map" with non-array schema value' => [
-                new RuleFactory(),
                 [
                     'type' => 'map',
                     'schema' => [
@@ -219,7 +201,6 @@ class ParserTest extends TestCase
                 'Invalid schema value: /schema/name/schema must be of type <array>, got <boolean>',
             ],
             'invalid schema value (nested)' => [
-                new RuleFactory(),
                 [
                     'type' => 'map',
                     'schema' => [
@@ -234,7 +215,6 @@ class ParserTest extends TestCase
                 'Invalid schema value: /schema/name/0 must be of type <array>, got <integer>',
             ],
             'unknown rule' => [
-                new RuleFactory(),
                 [
                     'type' => 'integer',
                     'whooops' => [1, 2, 3],
@@ -244,7 +224,6 @@ class ParserTest extends TestCase
                 'Unknown validation rule ID: /whooops specifies an unknown validation rule ID',
             ],
             'nested unknown rule' => [
-                new RuleFactory(),
                 [
                     'type' => 'list',
                     'schema' => [
@@ -257,7 +236,6 @@ class ParserTest extends TestCase
                 'Unknown validation rule ID: /schema/whooops specifies an unknown validation rule ID',
             ],
             'unknown rule with additions' => [
-                new RuleFactory(),
                 [
                     'type' => 'integer',
                     ['whooops' => [1, 2, 3]],
@@ -267,7 +245,6 @@ class ParserTest extends TestCase
                 'Unknown validation rule ID: /0/whooops specifies an unknown validation rule ID',
             ],
             'nested unknown rule with additions' => [
-                new RuleFactory(),
                 [
                     'type' => 'list',
                     'schema' => [
@@ -280,7 +257,6 @@ class ParserTest extends TestCase
                 'Unknown validation rule ID: /schema/0/whooops specifies an unknown validation rule ID',
             ],
             'rule with additions with missing rule' => [
-                new RuleFactory(),
                 [
                     'type' => 'string',
                     [],
@@ -290,7 +266,6 @@ class ParserTest extends TestCase
                 'Invalid schema: /0 contains no rule',
             ],
             'nested rule with additions with missing rule' => [
-                new RuleFactory(),
                 [
                     'type' => 'list',
                     'schema' => [
@@ -303,7 +278,6 @@ class ParserTest extends TestCase
                 'Invalid schema: /schema/0 contains no rule',
             ],
             'rule with additions with ambiguous rule' => [
-                new RuleFactory(),
                 [
                     'type' => 'string',
                     ['required' => false, 'blankable' => true],
@@ -313,7 +287,6 @@ class ParserTest extends TestCase
                 'Invalid schema: /0 contains multiple rules (["required","blankable"])',
             ],
             'nested rule with additions with ambiguous rule' => [
-                new RuleFactory(),
                 [
                     'type' => 'list',
                     'schema' => [
@@ -326,7 +299,6 @@ class ParserTest extends TestCase
                 'Invalid schema: /schema/0 contains multiple rules (["required","blankable"])',
             ],
             'invalid "message" data type' => [
-                new RuleFactory(),
                 [
                     'type' => 'boolean',
                     ['nullable' => true, 'message' => 13.37],
@@ -336,7 +308,6 @@ class ParserTest extends TestCase
                 'Invalid schema value: /0/message must be of type <string>, got <double>',
             ],
             'invalid "meta" data type' => [
-                new RuleFactory(),
                 [
                     'type' => 'float',
                     ['required' => true, 'meta' => 'foobar'],
@@ -346,7 +317,6 @@ class ParserTest extends TestCase
                 'Invalid schema value: /0/meta must be of type <array>, got <string>',
             ],
             'inapplicable rule' => [
-                new RuleFactory(),
                 [
                     'type' => 'integer',
                     'blankable' => true,
@@ -359,7 +329,6 @@ class ParserTest extends TestCase
     }
 
     /**
-     * @covers ::__construct
      * @covers ::ascertainInputHoldsArrayOrFail
      * @covers ::extractTypeOrFail
      * @covers ::fail
@@ -390,8 +359,7 @@ class ParserTest extends TestCase
             ],
         ]);
 
-        $ruleFactory = new RuleFactory();
-        $parser = new Parser($ruleFactory);
+        $parser = new Parser();
 
         $schema = $parser->parse([
             'type' => 'map',