+++ CHANGELOG.md
@@ -11,10 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - adds sanitization feature
 - introduces ```Ares\Ares``` as facade for data validation and sanitization
 - adds custom types to allow for easier schema reuse
+- adds ```Ares\Validation\RuleRegistry``` to replace ```Ares\Validation\RuleFactory```
 ### Changed
 - ```'required' => true``` is now validation default behavior 
 - renames validation option ```allowUnknown``` to ```allUnknownAllowed```
 - moves all validation related classes into sub-namespace ```Validation```
+### Removed
+- removed ```Ares\Validation\RuleFactory```
 
 ## [v1.1.1] 2019-04-20
 ### Added